% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aibw.R
\name{aibw}
\alias{aibw}
\title{Calculate Adjusted Ideal Body Weight}
\usage{
aibw(
  weight,
  height,
  sexf,
  age,
  allow_ibw_lt_intercept = TRUE,
  allow_tbw_lt_ibw = TRUE
)
}
\arguments{
\item{weight}{baseline weight of subject in kilograms}

\item{height}{baseline height of subject in centimeters}

\item{sexf}{0 = male, 1 = female, from sexf()}

\item{age}{Numeric vector of baseline age in years.}

\item{allow_ibw_lt_intercept}{logical indicating whether to allow ideal
body weight to be lower than intercepts (default TRUE).}

\item{allow_tbw_lt_ibw}{logical indicating whether to allow adjusted ideal
body weight to be less than IBW (default TRUE).}
}
\value{
adjusted ideal body weight (kg)
}
\description{
Calculates adjusted ideal body weight, which accounts for excess body weight
in obese patients.
}
\details{
Adjusted ideal body weight is calculated as:
\deqn{AIBW = IBW + 0.4 \cdot (TBW - IBW)}{AIBW = IBW + 0.4 * (TBW - IBW)}

where:
\itemize{
\item \eqn{IBW} = ideal body weight (kg)
\item \eqn{TBW} = total (actual) body weight (kg)
}
}
\examples{
df <- data.frame(
  ID = 1:6,
  HEIGHT = c(160, 170, 175, 165, 180, 150),
  WEIGHT = c(53, 71, 78, 55, 72, 43),
  SEX = c(1, 0, 0, 1, 0, 1),
  AGE = c(18, 27, 34, 33, 29, 30)
)
df <- dplyr::mutate(df, AIBW = aibw(WEIGHT, HEIGHT, SEX, AGE))
df
}
\seealso{
Other body_composition:
\code{\link[=bmi]{bmi()}},
\code{\link[=bmic]{bmic()}},
\code{\link[=bsa]{bsa()}},
\code{\link[=ibw]{ibw()}}
}
\concept{body_composition}
