% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bbmi}
\alias{bbmi}
\title{Calculate BMI (Deprecated)}
\usage{
bbmi(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=bmi]{bmi()}}}
}
\description{
\code{bbmi()} was renamed to \code{bmi()} to improve function naming consistency.
}
\keyword{internal}
