% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfc.R
\name{bhfc}
\alias{bhfc}
\title{Calculates hepatic function categories based on NCI-ODWG criteria}
\usage{
bhfc(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=hfc]{hfc()}}}
}
\description{
\code{bhfc()} was renamed to \code{hfc()} to improve function naming consistency.
}
\keyword{internal}
