% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfc.R
\name{brfc}
\alias{brfc}
\title{Renal Function Classification (Deprecated)}
\usage{
brfc(crcl)
}
\arguments{
\item{crcl}{Creatinine clearance value}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{brfc()} has been replaced by \code{rfc()} with a completely different interface.
}
\keyword{internal}
