% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{check_for_unique_units}
\alias{check_for_unique_units}
\title{Check for Unique Units per Parameter}
\usage{
check_for_unique_units(params, units)
}
\arguments{
\item{params}{a column from a dataset with lab parameters}

\item{units}{a column from a dataset with units associated with those parameters}
}
\value{
a boolean
}
\description{
Check for Unique Units per Parameter
}
\examples{
df <- data.frame(
  PARAM = c(
    "ALB","ALT","AST","CR","TBIL",
    "ALB","CR","TBIL","ALT","AST"),
  UNIT = c(
    "g/L","U/L","U/L","umol/L","umol/L",
    "U/L","μmol/L","μmol/L","IU/L","IU/L")
)
check_for_unique_units(df$PARAM, df$UNIT)
}
\seealso{
Other unit_checking:
\code{\link[=get_unique_units_df]{get_unique_units_df()}}
}
\concept{unit_checking}
