% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aegfr.R
\name{convert_abs_to_rel}
\alias{convert_abs_to_rel}
\title{Convert absolute eGFR to relative eGFR}
\usage{
convert_abs_to_rel(est, bsa)
}
\arguments{
\item{est}{Absolute eGFR (mL/min)}

\item{bsa}{Body surface area (m²)}
}
\value{
Relative eGFR (mL/min/1.73m²)
}
\description{
Convert absolute eGFR to relative eGFR
}
\keyword{internal}
