% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aegfr.R
\name{convert_rel_to_abs}
\alias{convert_rel_to_abs}
\title{Convert relative eGFR to absolute eGFR}
\usage{
convert_rel_to_abs(est, bsa)
}
\arguments{
\item{est}{Relative eGFR (mL/min/1.73m²)}

\item{bsa}{Body surface area (m²)}
}
\value{
Absolute eGFR (mL/min)
}
\description{
Convert relative eGFR to absolute eGFR
}
\keyword{internal}
