% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Creates the directory if it doesn't exist}
\usage{
create_dir(path)
}
\arguments{
\item{path}{path of directory to be created}
}
\value{
Nothing
}
\description{
Creates the directory if it doesn't exist
}
\examples{
\dontrun{
create_dir("derived/data/test")
}
}
