% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{.egfr_ckdepi_2021_cystatin}
\alias{.egfr_ckdepi_2021_cystatin}
\title{CKD-EPI 2021 creatinine-cystatin C eGFR equation}
\usage{
.egfr_ckdepi_2021_cystatin(sexf, age, creat, cystc)
}
\arguments{
\item{sexf}{a boolean representing if the patient is female.}

\item{age}{age of patient in years}

\item{creat}{serum creatinine levels in mg/dL.}

\item{cystc}{serum cystatin C levels in mg/L.}
}
\value{
eGFR in mL/min/1.73 m^2
}
\description{
The equation used by \code{\link[=egfr]{egfr()}} with \code{method = "CKDEPI 2021 cystatin"}.
}
\details{
The CKD-EPI 2021 creatinine-cystatin equation:
\deqn{eGFR = 135 \cdot \min(S_{cr}/\kappa, 1)^\alpha \cdot \max(S_{cr}/\kappa, 1)^{-0.544} \cdot \min(S_{cys}/0.8, 1)^{-0.323} \\
\cdot \max(S_{cys}/0.8, 1)^{-0.778} \cdot 0.9961^A \cdot 0.963^F}

where:
\itemize{
\item \eqn{S_{cr}} = serum creatinine (mg/dL)
\item \eqn{S_{cys}} = serum cystatin C (mg/L)
\item \eqn{\kappa} = 0.7 (female) or 0.9 (male)
\item \eqn{\alpha} = -0.219 (female) or -0.144 (male)
\item \eqn{A} = age (years)
\item \eqn{F} = 1 (female) or 0 (male)
}
}
\keyword{internal}
