% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{.egfr_mdrd}
\alias{.egfr_mdrd}
\title{MDRD (4-variable) eGFR equation}
\usage{
.egfr_mdrd(sexf, raceb, age, creat)
}
\arguments{
\item{sexf}{a boolean representing if the patient is female.}

\item{raceb}{a boolean representing if the patient is black.}

\item{age}{the age of the patient in years}

\item{creat}{the serum creatinine levels in mg/dL}
}
\value{
the eGFR in mL/min/1.73 m^2
}
\description{
The equation used by \code{\link[=egfr]{egfr()}} with \code{method = "MDRD"}.
}
\details{
The MDRD equation:
\deqn{eGFR = 175 \cdot S_{cr}^{-1.154} \cdot A^{-0.203} \cdot 0.742^F \cdot 1.212^B}{eGFR = 175 * Scr^-1.154 * A^-0.203 * 0.742^F * 1.212^B}

where:
\itemize{
\item \eqn{S_{cr}} = serum creatinine (mg/dL)
\item \eqn{A} = age (years)
\item \eqn{F} = 1 (female) or 0 (male)
\item \eqn{B} = 1 (Black) or 0 (other)
}
}
\keyword{internal}
