% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yspec_def.R
\name{ethnicn}
\alias{ethnicn}
\title{Convert Ethnicity to Numeric Code}
\usage{
ethnicn(ethnicc)
}
\arguments{
\item{ethnicc}{Ethnic character}
}
\value{
the standard yspec numeric value for the inputted Ethnic character
}
\description{
Convert Ethnicity to Numeric Code
}
\examples{
ethnicn("HISPANIC OR LATINO") # 1

ethnicn("NOT HISPANIC OR LATINO") # 0

ethnicn("UNKNOWN") # default missing value
}
\seealso{
Other demographics:
\code{\link[=is_american_native]{is_american_native()}},
\code{\link[=is_asian]{is_asian()}},
\code{\link[=is_black]{is_black()}},
\code{\link[=is_female]{is_female()}},
\code{\link[=is_hispanic_or_latino]{is_hispanic_or_latino()}},
\code{\link[=is_not_hispanic_or_latino]{is_not_hispanic_or_latino()}},
\code{\link[=is_other]{is_other()}},
\code{\link[=is_pacific_islander]{is_pacific_islander()}},
\code{\link[=is_unspecified]{is_unspecified()}},
\code{\link[=is_white]{is_white()}},
\code{\link[=racen]{racen()}},
\code{\link[=sexf]{sexf()}}
}
\concept{demographics}
