% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{geom_mean}
\alias{geom_mean}
\title{Calculate Geometric Mean}
\usage{
geom_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector to compute geometric mean of}

\item{na.rm}{boolean to remove NA from vector in calcualtion. Default is False}
}
\value{
geometric mean of input vector x
}
\description{
Calculate Geometric Mean
}
\details{
The geometric mean is calculated as:
\deqn{GM = \exp\left(\mathbb{E}[\log(x)]\right)}{GM = exp(mean(log(x)))}
}
\examples{
geom_mean(c(1, 2, 3, 2, 1))
}
\seealso{
Other statistics:
\code{\link[=cor_df]{cor_df()}},
\code{\link[=cv]{cv()}},
\code{\link[=geom_cv]{geom_cv()}},
\code{\link[=geom_sd]{geom_sd()}}
}
\concept{statistics}
