% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{geom_sd}
\alias{geom_sd}
\title{Calculate Geometric Standard Deviation}
\usage{
geom_sd(x, na.rm = FALSE)
}
\arguments{
\item{x}{The vector of data you want the geometric sd of.}

\item{na.rm}{a boolean to remove NA values. Default is False}
}
\value{
the geometric standard deviation of x
}
\description{
Calculate Geometric Standard Deviation
}
\details{
The geometric standard deviation is calculated as:
\deqn{GSD = \exp\left(\sqrt{Var[\log(x)]}\right)}{GSD = exp(sd(log(x)))}
}
\examples{
geom_sd(c(1, 2, 3, 2, 1))
}
\seealso{
Other statistics:
\code{\link[=cor_df]{cor_df()}},
\code{\link[=cv]{cv()}},
\code{\link[=geom_cv]{geom_cv()}},
\code{\link[=geom_mean]{geom_mean()}}
}
\concept{statistics}
