% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_asian}
\alias{is_asian}
\title{Check if Race is Asian}
\usage{
is_asian(x)
}
\arguments{
\item{x}{input character representing race}
}
\value{
boolean representing Race == Asian
}
\description{
Check if Race is Asian
}
\examples{
is_asian("ASIAN")

is_asian("BLACK")

is_asian(3)
}
\seealso{
Other demographics:
\code{\link[=ethnicn]{ethnicn()}},
\code{\link[=is_american_native]{is_american_native()}},
\code{\link[=is_black]{is_black()}},
\code{\link[=is_female]{is_female()}},
\code{\link[=is_hispanic_or_latino]{is_hispanic_or_latino()}},
\code{\link[=is_not_hispanic_or_latino]{is_not_hispanic_or_latino()}},
\code{\link[=is_other]{is_other()}},
\code{\link[=is_pacific_islander]{is_pacific_islander()}},
\code{\link[=is_unspecified]{is_unspecified()}},
\code{\link[=is_white]{is_white()}},
\code{\link[=racen]{racen()}},
\code{\link[=sexf]{sexf()}}
}
\concept{demographics}
