% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_other}
\alias{is_other}
\title{Check if Race is Other}
\usage{
is_other(x)
}
\arguments{
\item{x}{input character representing race}
}
\value{
boolean representing Race == Other
}
\description{
Check if Race is Other
}
\examples{
is_other("OTHER")

is_other("MULTIPLE")

is_other("BLACK")

is_other(6)
}
\seealso{
Other demographics:
\code{\link[=ethnicn]{ethnicn()}},
\code{\link[=is_american_native]{is_american_native()}},
\code{\link[=is_asian]{is_asian()}},
\code{\link[=is_black]{is_black()}},
\code{\link[=is_female]{is_female()}},
\code{\link[=is_hispanic_or_latino]{is_hispanic_or_latino()}},
\code{\link[=is_not_hispanic_or_latino]{is_not_hispanic_or_latino()}},
\code{\link[=is_pacific_islander]{is_pacific_islander()}},
\code{\link[=is_unspecified]{is_unspecified()}},
\code{\link[=is_white]{is_white()}},
\code{\link[=racen]{racen()}},
\code{\link[=sexf]{sexf()}}
}
\concept{demographics}
