% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_pacific_islander}
\alias{is_pacific_islander}
\title{Check if Race is Pacific Islander}
\usage{
is_pacific_islander(x)
}
\arguments{
\item{x}{input character representing race}
}
\value{
boolean representing Race == Pacific Islander
}
\description{
Check if Race is Pacific Islander
}
\examples{
is_pacific_islander("NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER")

is_pacific_islander("BLACK")

is_pacific_islander(5)
}
\seealso{
Other demographics: 
\code{\link{ethnicn}()},
\code{\link{is_american_native}()},
\code{\link{is_asian}()},
\code{\link{is_black}()},
\code{\link{is_female}()},
\code{\link{is_hispanic_or_latino}()},
\code{\link{is_not_hispanic_or_latino}()},
\code{\link{is_other}()},
\code{\link{is_unspecified}()},
\code{\link{is_white}()},
\code{\link{racen}()},
\code{\link{sexf}()}
}
\concept{demographics}
