% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{mosteller_bsa}
\alias{mosteller_bsa}
\title{Calculate Body Surface Area Using Mosteller Equation}
\usage{
mosteller_bsa(weight, height)
}
\arguments{
\item{weight}{weight of subject (kg)}

\item{height}{height of subject (cm)}
}
\value{
the body surface area (m^2)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated; use \code{\link[=bsa]{bsa()}} with \code{method = "Mosteller"}.
}
\details{
The Mosteller equation for BSA:
\deqn{BSA = \sqrt{\frac{W \cdot H}{3600}}}{BSA = sqrt(W * H / 3600)}

where:
\itemize{
\item \eqn{W} = weight (kg)
\item \eqn{H} = height (cm)
}
}
\examples{
mosteller_bsa <- bsa(70, 170, method = "Mosteller")
}
\keyword{internal}
