% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_with_hash.R
\name{read_excel_with_hash}
\alias{read_excel_with_hash}
\title{Read Excel File with Hash Verification}
\usage{
read_excel_with_hash(xlsx_file_path, ..., algo = "blake3")
}
\arguments{
\item{xlsx_file_path}{an xlsx/xls file to ingest}

\item{...}{additional arguments to digest or read_excel}

\item{algo}{hashing algorithm to use, default is "blake3"}
}
\value{
a dataframe(?) of data within file
}
\description{
Read Excel File with Hash Verification
}
\examples{
\dontrun{
read_excel_with_hash("data/source/example.xpt")
}
}
