% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_with_hash.R
\name{read_sas_with_hash}
\alias{read_sas_with_hash}
\title{Read SAS File with Hash Verification}
\usage{
read_sas_with_hash(sas_file_path, ..., algo = "blake3")
}
\arguments{
\item{sas_file_path}{path to sas file to ingest}

\item{...}{additional arguments to digest or read_sas}

\item{algo}{hashing algorithm to use, default is "blake3"}
}
\value{
a dataframe(?) of data within file
}
\description{
Read SAS File with Hash Verification
}
\examples{
\dontrun{
read_sas_with_hash("data/source/example.sas7bdat")
}
}
