% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yspec_def.R
\name{validate_racen_config}
\alias{validate_racen_config}
\title{Validate the scicalc.racen_config option}
\usage{
validate_racen_config(config)
}
\arguments{
\item{config}{the value of \code{getOption("scicalc.racen_config")}}
}
\value{
the validated config with lower-cased names, or \code{NULL}
}
\description{
Validate the scicalc.racen_config option
}
\keyword{internal}
