% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aegfr.R
\name{aegfr}
\alias{aegfr}
\title{Calculates Absolute estimated Glomerular Filtration Rate (eGFR)}
\usage{
aegfr(egfr, bsa)
}
\arguments{
\item{egfr}{estimated glomerular filtration rate (mL/min/1.73 m2)}

\item{bsa}{body surface area (m2)}
}
\value{
Absolute eGFR (mL/min)
}
\description{
Calculates Absolute estimated Glomerular Filtration Rate (eGFR)
}
\examples{
aegfr(90, 1.9)

df <- data.frame(
  ID = c(1, 2, 3, 4),
  EGFR = c(80, 95, 70, 60),
  BSA = c(1.60, 1.85, 1.75, 2.00)
)

df <- df \%>\%
  dplyr::group_by(ID) \%>\%
  dplyr::mutate(AEGFR = aegfr(EGFR, BSA))
df
}
