% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agec.R
\name{agec}
\alias{agec}
\title{Age category function}
\usage{
agec(age)
}
\arguments{
\item{age}{baseline age of subject in years}
}
\value{
age category
}
\description{
Age category function
}
\examples{
age_cat <- agec(24)

df <- data.frame(
  ID = 1:12,
  AGE = c(0.07, 28 / 365, 0.25, 1, 2, 4, 12, 16, 18, 24, 65, 70)
)
df <- dplyr::mutate(df, AGEC = agec(AGE))
df
}
\references{
Pediatric Drug Development: Regulatory Considerations —
Complying With the Pediatric Research Equity Act and
Qualifying for Pediatric Exclusivity Under the Best
Pharmaceuticals for Children Act Guidance for Industry

Guideline for Industry Studies in Support of
Special Populations: Geriatrics
}
