% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmi.R
\name{bbmi}
\alias{bbmi}
\title{Calculates Baseline Body Mass Index based on Weight and Height}
\usage{
bbmi(weight, height)
}
\arguments{
\item{weight}{weight of subject (kg)}

\item{height}{height of subject (cm)}
}
\value{
the bBMI value (kg m^(-2))
}
\description{
Calculates Baseline Body Mass Index based on Weight and Height
}
\examples{
b <- bbmi(80.56, 167)

df <- data.frame(
"WT" = c(80.56, 71.53, 81.04, 70.17),
"HT" = c(167, 161, 163, 164)
)
df <- dplyr::mutate(df, bbmi = bbmi(WT, HT))
}
