% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{Calculate Body Mass Index}
\usage{
bmi(weight, height)
}
\arguments{
\item{weight}{weight of subject (kg)}

\item{height}{height of subject (cm)}
}
\value{
the BMI value (kg/m^2)
}
\description{
Calculate Body Mass Index
}
\details{
BMI is calculated using the formula:
\deqn{BMI = \frac{W}{(H/100)^2}}{BMI = W / (H/100)^2}

where:
\itemize{
\item \eqn{W} = weight (kg)
\item \eqn{H} = height (cm)
}
}
\examples{
b <- bmi(80.56, 167)

df <- data.frame(
  "WT" = c(80.56, 71.53, 81.04, 70.17),
  "HT" = c(167, 161, 163, 164)
)
df <- dplyr::mutate(df, bmi = bmi(WT, HT))
df
}
\seealso{
Other body_composition: 
\code{\link{aibw}()},
\code{\link{bmic}()},
\code{\link{bsa}()},
\code{\link{ibw}()}
}
\concept{body_composition}
