% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{Calculates Baseline Body Mass Index based on Weight and Height}
\usage{
bmi(weight, height)
}
\arguments{
\item{weight}{weight of subject (kg)}

\item{height}{height of subject (cm)}
}
\value{
the BMI value (kg m^(-2))
}
\description{
Calculates Baseline Body Mass Index based on Weight and Height
}
\examples{
b <- bmi(80.56, 167)

df <- data.frame(
"WT" = c(80.56, 71.53, 81.04, 70.17),
"HT" = c(167, 161, 163, 164)
)
df <- dplyr::mutate(df, bmi = bmi(WT, HT))
df
}
