% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brfc.R
\name{brfc}
\alias{brfc}
\title{Calculates renal impairment categories based on Renal Function Estimator}
\usage{
brfc(
  est,
  est_units = c("mL/min", "mL/min/1.73m2"),
  bsa,
  category_standard = c("regulatory", "clinical")
)
}
\arguments{
\item{est}{Renal Function Estimator, either EGFR (mL/min/1.73m2), AEGFR (mL/min) or creatinine clearance rate (mL/min)}

\item{bsa}{Body surface are to normalize EGFR to AEGFR when calculating regulatory brfc}

\item{category_standard}{either regulatory or clinical}
}
\value{
integer renal impairment category
(Clinical category)
1: Normal, Estimator >= 90 mL/min/1.73m2
2: Mild, 90 > Estimator >=60
3: Moderate, 60 > Estimator >= 30
4: Severe, 30 > Estimator >= 15
5: End Stage, Estimator < 15

(Regulatory category)
1: Normal, Estimator >= 90 mL/min
2: Mild, 90 > Estimator >=60
3: Moderate, 60 > Estimator >= 30
4: Severe, Estimator < 30
}
\description{
Calculates renal impairment categories based on Renal Function Estimator
}
\examples{
brfc(crcl(FALSE, 20, 10, 70))

df <- data.frame(
  "ID" = c(1, 1, 1, 1, 2, 2, 2, 2),
  "SEX" = c("F", "F", "F", "F", "M", "M", "M", "M"),
  "RACE" = c("WHITE", "WHITE", "WHITE", "WHITE", "BLACK", "BLACK", "BLACK", "BLACK"),
  "AGE" = c(24, 24, 24, 24, 22, 22, 22, 22),
  "CREAT" = c(1, 1, 1, 1, 4, 4, 4, 4),
  "WEIGHT" = c(70, 70, 70, 70, 65, 65, 65, 65)
)

df <- df \%>\%
  dplyr::group_by(ID) \%>\%
  dplyr::mutate(
    CRCL = crcl(is_female(SEX), AGE, CREAT, WEIGHT),
    BRFC = brfc(CRCL)
  )
}
