% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{bsa}
\alias{bsa}
\title{Calculate Body Surface Area}
\usage{
bsa(weight, height, method = "Dubois")
}
\arguments{
\item{weight}{weight of a subject (kg)}

\item{height}{height of a subject (cm)}

\item{method}{equation to use: \link[=.bsa_dubois]{bsa(method = "Dubois")} or
\link[=.bsa_mosteller]{bsa(method = "Mosteller")}.}
}
\value{
bsa (m^2)
}
\description{
Calculate Body Surface Area
}
\examples{
bsa(70, 170)
bsa(70, 170, method = "Mosteller")
bsa(70, 170, method = "Dubois")
}
\seealso{
The method implementations:
\link[=.bsa_dubois]{bsa(method = "Dubois")},
\link[=.bsa_mosteller]{bsa(method = "Mosteller")}.

Other body_composition: 
\code{\link{aibw}()},
\code{\link{bmi}()},
\code{\link{bmic}()},
\code{\link{ibw}()}
}
\concept{body_composition}
