% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{bsa}
\alias{bsa}
\title{Calculate Body Surface Area}
\usage{
bsa(weight, height, method = "Dubois")
}
\arguments{
\item{weight}{weight of a subject (kg)}

\item{height}{height of a subject (cm)}

\item{method}{String to dictate which equation to use. Dubois or Mosteller.}
}
\value{
bsa (m^2)
}
\description{
Calculate Body Surface Area
}
\examples{
bsa(70, 170)
bsa(70, 170, method = "Mosteller")
bsa(70, 170, method = "Dubois")
}
\seealso{
Other body_composition: 
\code{\link{aibw}()},
\code{\link{bmi}()},
\code{\link{bmic}()},
\code{\link{dubois_bsa}()},
\code{\link{ibw}()},
\code{\link{mosteller_bsa}()}
}
\concept{body_composition}
