% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Categorize Continuous Variable into Quantile Bins}
\usage{
categorize(continuous_var, nbins = 4, units = "", type = 7, digits = 1)
}
\arguments{
\item{continuous_var}{continuous variable data}

\item{nbins}{number of bins to break data into, default is 4}

\item{units}{string, optional units string to add to labels of categorized data}

\item{type}{type argument for stats::quantile, default is 7}

\item{digits}{number of digits to round quantile breaks to for labels, default is 1}
}
\value{
a vector of categorized data as factor
}
\description{
Categorize Continuous Variable into Quantile Bins
}
\examples{
x <- rnorm(1000, mean = 10, sd = 5)
categorize(x, nbins = 5)
}
