% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{check_for_unique_units}
\alias{check_for_unique_units}
\title{Gives a TRUE/FALSE for if the Parameters have only 1 associated unit}
\usage{
check_for_unique_units(params, units)
}
\arguments{
\item{params}{a column from a dataset with lab parameters}

\item{units}{a column from a dataset with units associated with those parameters}
}
\value{
a boolean
}
\description{
Gives a TRUE/FALSE for if the Parameters have only 1 associated unit
}
\examples{
df <- data.frame(
  PARAM = c(
    "ALB","ALT","AST","CR","TBIL",
    "ALB","CR","TBIL","ALT","AST"),
  UNIT = c(
    "g/L","U/L","U/L","umol/L","umol/L",
    "U/L","μmol/L","μmol/L","IU/L","IU/L")
)
check_for_unique_units <- get_unique_units_df(df$PARAM, df$UNIT)
}
