% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{ckdepi_2009_egfr}
\alias{ckdepi_2009_egfr}
\title{Calculates Estimated Glomerular Filtration Rate based on Sex, Race, Age, and Creatinine levels
based on the CKDEPI 2009 equation}
\usage{
ckdepi_2009_egfr(sexf, raceb, age, creat)
}
\arguments{
\item{sexf}{boolean value of sex Female: TRUE, Male: FALSE}

\item{raceb}{boolean value of Race == Black: Black: TRUE, Other: FALSE}

\item{age}{age of subject (years)}

\item{creat}{creatinine levels of subject (mg/dL)}
}
\value{
the eGFR value (mL/min/1.73m2)
}
\description{
Calculates Estimated Glomerular Filtration Rate based on Sex, Race, Age, and Creatinine levels
based on the CKDEPI 2009 equation
}
\examples{
e <- ckdepi_2009_egfr(TRUE, TRUE, 24, 1)

df <- data.frame(
   "SEXF" = c(TRUE, FALSE, TRUE, FALSE),
   "RACEB" = c(FALSE, FALSE, TRUE, FALSE),
   "AGE" = c(24, 24, 23, 24),
   "CREAT" = c(1, 1, 2, 1)
   )
df <- dplyr::mutate(df, egfr = ckdepi_2009_egfr(SEXF, RACEB, AGE, CREAT))
}
