% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{ckdepi_2021_egfr}
\alias{ckdepi_2021_egfr}
\title{Calculates eGFR using the CKDEPI 2021 creatinine equation}
\usage{
ckdepi_2021_egfr(sexf, age, creat)
}
\arguments{
\item{sexf}{boolean value of sex Female: TRUE, Male: FALSE}

\item{age}{age of subject (years)}

\item{creat}{creatinine levels of subject (mg/dL)}
}
\value{
the eGFR value (mL/min/1.73m2)
}
\description{
Calculates eGFR using the CKDEPI 2021 creatinine equation
}
\examples{
e <- ckdepi_2021_egfr(TRUE, 24, 1)

df <- data.frame(
   "SEXF" = c(TRUE, FALSE, TRUE, FALSE),
   "RACEB" = c(FALSE, FALSE, TRUE, FALSE),
   "AGE" = c(24, 24, 23, 24),
   "CREAT" = c(1, 1, 2, 1)
   )
df <- dplyr::mutate(df, egfr = ckdepi_2021_egfr(SEXF, AGE, CREAT))
}
