% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{ckdepi_2021_egfr}
\alias{ckdepi_2021_egfr}
\title{Calculates eGFR with CKDEPI 2021 equation}
\usage{
ckdepi_2021_egfr(sexf, age, creat, cystc)
}
\arguments{
\item{sexf}{a boolean representing if the patient is female.}

\item{age}{age of patient in years}

\item{creat}{serum creatinine levels in mg/dL.}

\item{cystc}{serum cystain C levels in mg/L.}
}
\value{
eGFR in mL/min/1.73 m^2
}
\description{
Calculates eGFR with CKDEPI 2021 equation
}
\examples{
e <- ckdepi_2021_egfr(TRUE, 24, 1, 2)

df <- data.frame(
   "SEXF" = c(TRUE, FALSE, TRUE, FALSE),
   "RACEB" = c(FALSE, FALSE, TRUE, FALSE),
   "AGE" = c(24, 24, 23, 24),
   "CREAT" = c(1, 1, 2, 1),
   "CYSTC" = c(0.4, 0.8, 1, 2)
   )
df <- dplyr::mutate(df, egfr = ckdepi_2021_egfr(SEXF, AGE, CREAT, CYSTC))
}
