% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_alb}
\alias{convert_alb}
\title{Convert Albumin concentration from g/L to g/dL}
\usage{
convert_alb(alb)
}
\arguments{
\item{alb}{albumin concentration (g/L)}
}
\value{
Albumin concentration (g/dL)
}
\description{
Convert Albumin concentration from g/L to g/dL
}
\examples{
convert_alb(40)

df <- data.frame(
  ID = c(1, 2, 3, 4),
  ALB = c(35, 40, 28, 45)
)

df <- df \%>\%
  dplyr::group_by(ID) \%>\%
  dplyr::mutate(ALBBL = convert_alb(ALB))
df
}
