% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_bili}
\alias{convert_bili}
\title{Convert Bilirubin Concentration Units}
\usage{
convert_bili(bili)
}
\arguments{
\item{bili}{bilirubin concentration (µmol/L)}
}
\value{
Bilirubin concentration (mg/dL)
}
\description{
Convert Bilirubin Concentration Units
}
\examples{
convert_bili(17.1) # ≈ 1 mg/dL

df <- data.frame(
  ID = c(1, 2, 3, 4),
  BILI = c(10, 15, 25, 40)
)

df <- df \%>\%
  dplyr::group_by(ID) \%>\%
  dplyr::mutate(BILIBL = convert_bili(BILI))
df
}
\seealso{
Other unit_conversion: 
\code{\link{convert_alb}()},
\code{\link{convert_creat}()},
\code{\link{convert_mass_to_mol}()},
\code{\link{convert_mol_to_mass}()}
}
\concept{unit_conversion}
