% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_bili}
\alias{convert_bili}
\title{Convert Bilirubin concentration from µmol/L to mg/dL}
\usage{
convert_bili(bili)
}
\arguments{
\item{bili}{bilirubin concentration (µmol/L)}
}
\value{
Bilirubin concentration (mg/dL)
}
\description{
Convert Bilirubin concentration from µmol/L to mg/dL
}
\examples{
convert_bili(17.1) # ≈ 1 mg/dL

df <- data.frame(
  ID = c(1, 2, 3, 4),
  BILI = c(10, 15, 25, 40)
)

df <- df \%>\%
  dplyr::group_by(ID) \%>\%
  dplyr::mutate(BILIBL = convert_bili(BILI))
df
}
