% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_creat}
\alias{convert_creat}
\title{Convert Serum Creatinine Concentration Units}
\usage{
convert_creat(creat)
}
\arguments{
\item{creat}{serum creatinine concentration (µmol/L)}
}
\value{
Serum Creatinine concentration (mg/dL)
}
\description{
Convert Serum Creatinine Concentration Units
}
\examples{
convert_creat(88.42) # ≈ 1 mg/dL

df <- data.frame(
  ID = c(1, 2, 3, 4),
  CREAT = c(70, 90, 110, 130)
)

df <- df |>
  dplyr::group_by(ID) |>
  dplyr::mutate(CREATBL = convert_creat(CREAT))
df
}
\seealso{
Other unit_conversion: 
\code{\link{convert_alb}()},
\code{\link{convert_bili}()}
}
\concept{unit_conversion}
