% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcl.R
\name{crcl}
\alias{crcl}
\title{Calculate Creatinine Clearance Using Cockcroft-Gault Equation}
\usage{
crcl(sexf, age, creat, weight)
}
\arguments{
\item{sexf}{bool of sex of subject. Female: True, Male: False}

\item{age}{age of subject (years)}

\item{creat}{serum creatinine levels (mg/dL)}

\item{weight}{weight of subject (kg)}
}
\value{
CrCl (mL/min)
}
\description{
Calculate Creatinine Clearance Using Cockcroft-Gault Equation
}
\details{
The Cockcroft-Gault equation:
\deqn{CrCl = F \cdot \frac{(140 - A) \cdot W}{72 \cdot S_{cr}}}{CrCl = F * (140 - A) * W / (72 * Scr)}

where:
\itemize{
\item \eqn{A} = age (years)
\item \eqn{W} = weight (kg)
\item \eqn{S_{cr}} = serum creatinine (mg/dL)
\item \eqn{F} = 0.85 (female) or 1 (male)
}
}
\examples{
crcl(FALSE, 20, 10, 70)

df <- data.frame(
  "ID" = c(1, 1, 1, 1, 2, 2, 2, 2),
  "SEX" = c("F", "F", "F", "F", "M", "M", "M", "M"),
  "RACE" = c("WHITE", "WHITE", "WHITE", "WHITE", "BLACK", "BLACK", "BLACK", "BLACK"),
  "AGE" = c(24, 24, 24, 24, 22, 22, 22, 22),
  "CREAT" = c(1, 1, 1, 1, 4, 4, 4, 4),
  "WEIGHT" = c(70, 70, 70, 70, 65, 65, 65, 65)
)

df <- df \%>\%
  dplyr::group_by(ID) \%>\%
  dplyr::mutate(CRCL = crcl(is_female(SEX), AGE, CREAT, WEIGHT))
df
}
\seealso{
Other renal_function: 
\code{\link{aegfr}()},
\code{\link{ckdepi_2009_egfr}()},
\code{\link{ckdepi_2021_egfr}()},
\code{\link{ckdepi_2021_egfr_cystatin}()},
\code{\link{egfr}()},
\code{\link{mdrd_egfr}()},
\code{\link{rfc}()},
\code{\link{schwartz_egfr}()}
}
\concept{renal_function}
