% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcl.R
\name{crcl}
\alias{crcl}
\title{Calculates Creatinine clearance with Cockcroft-Gault equation}
\usage{
crcl(sexf, age, creat, weight)
}
\arguments{
\item{sexf}{bool of sex of subject. Female: True, Male: False}

\item{age}{age of subject (years)}

\item{creat}{serum creatinine levels (mg/dL)}

\item{weight}{weight of subject (kg)}
}
\value{
CrCl (mL/min)
}
\description{
Calculates Creatinine clearance with Cockcroft-Gault equation
}
\examples{
crcl(FALSE, 20, 10, 70)

df <- data.frame(
  "ID" = c(1, 1, 1, 1, 2, 2, 2, 2),
  "SEX" = c("F", "F", "F", "F", "M", "M", "M", "M"),
  "RACE" = c("WHITE", "WHITE", "WHITE", "WHITE", "BLACK", "BLACK", "BLACK", "BLACK"),
  "AGE" = c(24, 24, 24, 24, 22, 22, 22, 22),
  "CREAT" = c(1, 1, 1, 1, 4, 4, 4, 4),
  "WEIGHT" = c(70, 70, 70, 70, 65, 65, 65, 65)
)

df <- df \%>\%
  dplyr::group_by(ID) \%>\%
  dplyr::mutate(CRCL = crcl(is_female(SEX), AGE, CREAT, WEIGHT))
}
