% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cv}
\alias{cv}
\title{Calculate Coefficient of Variation}
\usage{
cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{Input vector to compute CV for.}

\item{na.rm}{boolean to remove NA. default is FALSE}
}
\value{
CV of x. Standard deviation divided by mean. If you want \% you'll need to multiply by 100
}
\description{
Calculate Coefficient of Variation
}
\details{
The coefficient of variation is calculated as:
\deqn{CV = \frac{\sigma}{\mu}}{CV = sd / mean}

where:
\itemize{
\item \eqn{\sigma} = standard deviation
\item \eqn{\mu} = mean
}
}
\examples{
cv(c(1, 2, 1, 1, 2, 1, 2, 3))
}
\seealso{
Other statistics: 
\code{\link{cor_df}()},
\code{\link{geom_cv}()},
\code{\link{geom_mean}()},
\code{\link{geom_sd}()}
}
\concept{statistics}
