% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cv}
\alias{cv}
\title{Computes the coefficient of variation of input vector.}
\usage{
cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{Input vector to compute CV for.}

\item{na.rm}{boolean to remove NA. default is FALSE}
}
\value{
CV of x. Standard deviation divided by mean. If you want \% you'll need to multiply by 100
}
\description{
Computes the coefficient of variation of input vector.
}
\examples{
cv(c(1, 2, 1, 1, 2, 1, 2, 3))
}
