% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{.bsa_dubois}
\alias{.bsa_dubois}
\title{Du Bois body surface area equation}
\description{
The equation used by \code{\link[=bsa]{bsa()}} with \code{method = "Dubois"}.
}
\details{
The Du Bois equation for BSA:
\deqn{BSA = 0.007184 \cdot W^{0.425} \cdot H^{0.725}}{BSA = 0.007184 * W^0.425 * H^0.725}

where:
\itemize{
\item \eqn{W} = weight (kg)
\item \eqn{H} = height (cm)
}
}
\keyword{internal}
