% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{.bsa_mosteller}
\alias{.bsa_mosteller}
\title{Mosteller body surface area equation}
\description{
The equation used by \code{\link[=bsa]{bsa()}} with \code{method = "Mosteller"}.
}
\details{
The Mosteller equation for BSA:
\deqn{BSA = \sqrt{\frac{W \cdot H}{3600}}}{BSA = sqrt(W * H / 3600)}

where:
\itemize{
\item \eqn{W} = weight (kg)
\item \eqn{H} = height (cm)
}
}
\keyword{internal}
