% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{.egfr_schwartz}
\alias{.egfr_schwartz}
\title{Bedside Schwartz eGFR equation}
\usage{
.egfr_schwartz(height, creat)
}
\arguments{
\item{height}{height of patients in cm.}

\item{creat}{Serum creatinine levels in mg/dL}
}
\value{
eGFR in mL/min/1.73m^2
}
\description{
The equation used by \code{\link[=egfr]{egfr()}} with \code{method = "Schwartz"}.
}
\details{
The Schwartz equation for pediatric eGFR:
\deqn{eGFR = 0.413 \cdot \frac{H}{S_{cr}}}{eGFR = 0.413 * H / Scr}

where:
\itemize{
\item \eqn{H} = height (cm)
\item \eqn{S_{cr}} = serum creatinine (mg/dL)
}
}
\keyword{internal}
