% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{dubois_bsa}
\alias{dubois_bsa}
\title{Calculate Body Surface Area Using Du Bois Equation}
\usage{
dubois_bsa(weight, height)
}
\arguments{
\item{weight}{weight of subject (kg)}

\item{height}{height of subject (cm)}
}
\value{
the body surface area (m^2)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated; use \code{\link[=bsa]{bsa()}} with \code{method = "Dubois"}.
}
\details{
The Du Bois equation for BSA:
\deqn{BSA = 0.007184 \cdot W^{0.425} \cdot H^{0.725}}{BSA = 0.007184 * W^0.425 * H^0.725}

where:
\itemize{
\item \eqn{W} = weight (kg)
\item \eqn{H} = height (cm)
}
}
\examples{
b <- bsa(80.56, 167, method = "Dubois")
}
\keyword{internal}
