% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{dubois_bsa}
\alias{dubois_bsa}
\title{Calculates Body Surface Area based on Weight and Height using Dubois Dubois equation}
\usage{
dubois_bsa(weight, height)
}
\arguments{
\item{weight}{weight of subject (kg)}

\item{height}{height of subject (cm)}
}
\value{
the body surface area (m^2)
}
\description{
Calculates Body Surface Area based on Weight and Height using Dubois Dubois equation
}
\examples{
#' b <- dubois_bsa(80.56, 167)

df <- data.frame(
"WT" = c(80.56, 71.53, 81.04, 70.17),
"HT" = c(167, 161, 163, 164)
)
df <- dplyr::mutate(df, bsa = dubois_bsa(WT, HT))
}
