% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{egfr}
\alias{egfr}
\title{Calculate Estimated Glomerular Filtration Rate}
\usage{
egfr(
  sexf,
  raceb,
  age,
  creat,
  ...,
  cystc = NULL,
  height = NULL,
  method = "CKDEPI 2021"
)
}
\arguments{
\item{sexf}{a boolean representing if the patient is female.}

\item{raceb}{a boolean representing if the patient is black.}

\item{age}{the age of a patient in years.}

\item{creat}{the serum creatinine levels in mg/dL.}

\item{...}{reserved; must be empty. Arguments after \code{creat} must be named.}

\item{cystc}{the cystatin C levels in mg/L - required for the "CKDEPI 2021 cystatin" method.}

\item{height}{the height of a patient in cm - required for the "Schwartz" method.}

\item{method}{equation to use, one of:
\link[=.egfr_ckdepi_2009]{egfr(method = "CKDEPI 2009")},
\link[=.egfr_mdrd]{egfr(method = "MDRD")},
\link[=.egfr_ckdepi_2021]{egfr(method = "CKDEPI 2021")},
\link[=.egfr_ckdepi_2021_cystatin]{egfr(method = "CKDEPI 2021 cystatin")},
\link[=.egfr_schwartz]{egfr(method = "Schwartz")}.}
}
\value{
the eGFR calculated based on method.
}
\description{
Calculate Estimated Glomerular Filtration Rate
}
\examples{
e <- egfr(TRUE, TRUE, 24, 1, method = "CKDEPI 2009")

df <- data.frame(
  "SEXF" = c(TRUE, FALSE, TRUE, FALSE),
  "RACEB" = c(FALSE, FALSE, TRUE, FALSE),
  "AGE" = c(24, 24, 23, 24),
  "CREAT" = c(1, 1, 2, 1)
)
df <- dplyr::mutate(df, egfr = egfr(SEXF, RACEB, AGE, CREAT, method = "CKDEPI 2009"))
df
}
\seealso{
The method implementations:
\link[=.egfr_ckdepi_2009]{egfr(method = "CKDEPI 2009")},
\link[=.egfr_mdrd]{egfr(method = "MDRD")},
\link[=.egfr_ckdepi_2021]{egfr(method = "CKDEPI 2021")},
\link[=.egfr_ckdepi_2021_cystatin]{egfr(method = "CKDEPI 2021 cystatin")},
\link[=.egfr_schwartz]{egfr(method = "Schwartz")}.

Other renal_function: 
\code{\link{aegfr}()},
\code{\link{crcl}()},
\code{\link{rfc}()}
}
\concept{renal_function}
