% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{egfr}
\alias{egfr}
\title{Calculate Estimated Glomerular Filtration Rate}
\usage{
egfr(sexf, raceb, age, creat, cystc, height, method = "CKDEPI 2021")
}
\arguments{
\item{sexf}{a boolean representing if the patient is female.}

\item{raceb}{a boolean representing if the patient is black.}

\item{age}{the age of a patient in years.}

\item{creat}{the serum creatinine levels in mg/dL.}

\item{cystc}{the cystatin C levels in mg/L - only used in CKDEPI 2021 cystatin method}

\item{height}{the height of a patient in cm.}

\item{method}{a string specifying the method to use. Available options are "CKDEPI 2009", "MDRD", "CKDEPI 2021", "Schwartz".}
}
\value{
the eGFR calculated based on method.
}
\description{
Calculate Estimated Glomerular Filtration Rate
}
\examples{
e <- egfr(TRUE, TRUE, 24, 1, "CKDEPI 2009")

df <- data.frame(
  "SEXF" = c(TRUE, FALSE, TRUE, FALSE),
  "RACEB" = c(FALSE, FALSE, TRUE, FALSE),
  "AGE" = c(24, 24, 23, 24),
  "CREAT" = c(1, 1, 2, 1)
)
df <- dplyr::mutate(df, egfr = egfr(SEXF, RACEB, AGE, CREAT, "CKDEPI 2009"))
df
}
\seealso{
Other renal_function: 
\code{\link{aegfr}()},
\code{\link{ckdepi_2009_egfr}()},
\code{\link{ckdepi_2021_egfr}()},
\code{\link{ckdepi_2021_egfr_cystatin}()},
\code{\link{crcl}()},
\code{\link{mdrd_egfr}()},
\code{\link{rfc}()},
\code{\link{schwartz_egfr}()}
}
\concept{renal_function}
