% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{egfr}
\alias{egfr}
\title{Calculates eGFR based on the method specified}
\usage{
egfr(sexf, raceb, age, creat, cystc, height, method = "CKDEPI 2009")
}
\arguments{
\item{sexf}{a boolean representing if the patient is female.}

\item{raceb}{a boolean representing if the patient is black.}

\item{age}{the age of a patient in years.}

\item{creat}{the serum creatinine levels in mg/dL.}

\item{cystc}{the cystain C levels in mg/L - only used in CKDEPI 2021 method}

\item{height}{the height of a patient in cm.}

\item{method}{a string specifying the method to use. Available options are "CKDEPI 2009", "MDRD", "CKDEPI 2021", "Schwartz".}
}
\value{
the eGFR calculated based on method.
}
\description{
Calculates eGFR based on the method specified
}
\examples{
e <- egfr(TRUE, TRUE, 24, 1, "CKDEPI 2009")

df <- data.frame(
   "SEXF" = c(TRUE, FALSE, TRUE, FALSE),
   "RACEB" = c(FALSE, FALSE, TRUE, FALSE),
   "AGE" = c(24, 24, 23, 24),
   "CREAT" = c(1, 1, 2, 1)
   )
df <- dplyr::mutate(df, egfr = egfr(SEXF, RACEB, AGE, CREAT, "CKDEPI 2009"))
}
