% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yspec_def.R
\name{ethnicn}
\alias{ethnicn}
\title{Takes character input and returns standard yspec numeric value for Ethnic}
\usage{
ethnicn(ethnicc)
}
\arguments{
\item{ethnicc}{Ethnic character}
}
\value{
the standard yspec numeric value for the inputted Ethnic character
}
\description{
Takes character input and returns standard yspec numeric value for Ethnic
}
\examples{
ethnicn("HISPANIC OR LATINO") # 1

ethnicn("NOT HISPANIC OR LATINO") # 0

ethnicn("UNKNOWN") # -999
}
