% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{geom_cv}
\alias{geom_cv}
\title{Calculate Geometric Coefficient of Variation}
\usage{
geom_cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector of data you want the geometric CV of.}

\item{na.rm}{boolean to remove NA from vector. Default is FALSE}
}
\value{
the geometric CV of the input vector x
}
\description{
Calculate Geometric Coefficient of Variation
}
\details{
The geometric coefficient of variation is calculated as:
\deqn{GCV = \sqrt{\exp(Var[\log(x)]) - 1}}{GCV = sqrt(exp(sd(log(x))^2) - 1)}
}
\examples{
geom_cv(c(1, 2, 3, 2, 1))
}
\seealso{
Other statistics: 
\code{\link{cor_df}()},
\code{\link{cv}()},
\code{\link{geom_mean}()},
\code{\link{geom_sd}()}
}
\concept{statistics}
