% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{geom_cv}
\alias{geom_cv}
\title{Computes the geometric CV of a vector x}
\usage{
geom_cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector of data you want the geometric CV of.}

\item{na.rm}{boolean to remove NA from vector. Default is FALSE}
}
\value{
the geometric CV of the input vector x
}
\description{
Computes the geometric CV of a vector x
}
\examples{
geom_cv(c(1, 2, 3, 2, 1))
}
