% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{geom_mean}
\alias{geom_mean}
\title{Computes the geometric mean of a vector.}
\usage{
geom_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector to compute geometric mean of}

\item{na.rm}{boolean to remove NA from vector in calcualtion. Default is False}
}
\value{
geometric mean of input vector x
}
\description{
Computes the geometric mean of a vector.
}
\examples{
geom_mean(c(1, 2, 3, 2, 1))
}
