% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{geom_sd}
\alias{geom_sd}
\title{Computes the geometric standard deviation of a vector x.}
\usage{
geom_sd(x, na.rm = FALSE)
}
\arguments{
\item{x}{The vector of data you want the geometric sd of.}

\item{na.rm}{a boolean to remove NA values. Default is False}
}
\value{
the geometric standard deviation of x
}
\description{
Computes the geometric standard deviation of a vector x.
}
\examples{
geom_sd(c(1, 2, 3, 2, 1))
}
