% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_unique_units_df}
\alias{get_unique_units_df}
\title{Get Unique Parameter-Unit Combinations}
\usage{
get_unique_units_df(params, units)
}
\arguments{
\item{params}{a column from a dataset with lab parameters}

\item{units}{a column from a dataset with units associated with those parameters}
}
\value{
a dataframe with distinct units and parameters with IU replaced to U and mu replaced with u
}
\description{
Get Unique Parameter-Unit Combinations
}
\examples{
df <- data.frame(
  PARAM = c(
    "ALB","ALT","AST","CR","TBIL",
    "ALB","CR","TBIL","ALT","AST"),
  UNIT = c(
    "g/L","U/L","U/L","umol/L","umol/L",
    "U/L","μmol/L","μmol/L","IU/L","IU/L")
)
get_unique_units_df(df$PARAM, df$UNIT)
}
\seealso{
Other unit_checking: 
\code{\link{check_for_unique_units}()},
\code{\link{convert_units_to_spec}()},
\code{\link{pivot_with_units}()}
}
\concept{unit_checking}
