% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_american_native}
\alias{is_american_native}
\title{Check if Race is American Native}
\usage{
is_american_native(x)
}
\arguments{
\item{x}{input character representing race}
}
\value{
boolean representing Race == American Native
}
\description{
Check if Race is American Native
}
\examples{
is_american_native("AMERICAN INDIAN OR ALASKA NATIVE")

is_american_native("BLACK")

is_american_native(4)
}
\seealso{
Other demographics: 
\code{\link{ethnicn}()},
\code{\link{is_asian}()},
\code{\link{is_black}()},
\code{\link{is_female}()},
\code{\link{is_hispanic_or_latino}()},
\code{\link{is_not_hispanic_or_latino}()},
\code{\link{is_other}()},
\code{\link{is_pacific_islander}()},
\code{\link{is_unspecified}()},
\code{\link{is_white}()},
\code{\link{racen}()},
\code{\link{sexf}()}
}
\concept{demographics}
