% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_asian}
\alias{is_asian}
\title{Check if Race is Asian}
\usage{
is_asian(x)
}
\arguments{
\item{x}{input character representing race}
}
\value{
boolean representing Race == Asian
}
\description{
Check if Race is Asian
}
\examples{
is_asian("ASIAN")

is_asian("BLACK")

is_asian(3)
}
\seealso{
Other demographics: 
\code{\link{ethnicn}()},
\code{\link{is_american_native}()},
\code{\link{is_black}()},
\code{\link{is_female}()},
\code{\link{is_hispanic_or_latino}()},
\code{\link{is_not_hispanic_or_latino}()},
\code{\link{is_other}()},
\code{\link{is_pacific_islander}()},
\code{\link{is_unspecified}()},
\code{\link{is_white}()},
\code{\link{racen}()},
\code{\link{sexf}()}
}
\concept{demographics}
