% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_asian}
\alias{is_asian}
\title{Takes character input and returns TRUE/FALSE if asian/other}
\usage{
is_asian(x)
}
\arguments{
\item{x}{input character representing race}
}
\value{
boolean representing Race == Asian
}
\description{
Takes character input and returns TRUE/FALSE if asian/other
}
\examples{
is_asian("ASIAN")

is_asian("BLACK")
}
