% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_black}
\alias{is_black}
\title{Check if Race is Black}
\usage{
is_black(x)
}
\arguments{
\item{x}{input character representing race}
}
\value{
boolean representing Race == Black
}
\description{
Check if Race is Black
}
\examples{
is_black("WHITE")

is_black(c("AFRICAN AMERICAN", "BLACK"))

is_black(2)
}
\seealso{
Other demographics:
\code{\link[=ethnicn]{ethnicn()}},
\code{\link[=is_american_native]{is_american_native()}},
\code{\link[=is_asian]{is_asian()}},
\code{\link[=is_female]{is_female()}},
\code{\link[=is_hispanic_or_latino]{is_hispanic_or_latino()}},
\code{\link[=is_not_hispanic_or_latino]{is_not_hispanic_or_latino()}},
\code{\link[=is_other]{is_other()}},
\code{\link[=is_pacific_islander]{is_pacific_islander()}},
\code{\link[=is_unspecified]{is_unspecified()}},
\code{\link[=is_white]{is_white()}},
\code{\link[=racen]{racen()}},
\code{\link[=sexf]{sexf()}}
}
\concept{demographics}
