% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_black}
\alias{is_black}
\title{Check if Race is Black}
\usage{
is_black(x)
}
\arguments{
\item{x}{input character representing race}
}
\value{
boolean representing Race == Black
}
\description{
Check if Race is Black
}
\examples{
is_black("WHITE")

is_black(c("AFRICAN AMERICAN", "BLACK"))

is_black(2)
}
\seealso{
Other demographics: 
\code{\link{ethnicn}()},
\code{\link{is_asian}()},
\code{\link{is_female}()},
\code{\link{is_hispanic_or_latino}()},
\code{\link{is_not_hispanic_or_latino}()},
\code{\link{is_other}()},
\code{\link{is_white}()},
\code{\link{racen}()},
\code{\link{sexf}()}
}
\concept{demographics}
