% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_female}
\alias{is_female}
\title{Check if Sex is Female}
\usage{
is_female(x)
}
\arguments{
\item{x}{input character representing female or male}
}
\value{
boolean representing female
}
\description{
Check if Sex is Female
}
\examples{
is_female("F")

is_female(c("MALE", "FEMALE"))

is_female(c(1, 0, -999))
}
\seealso{
Other demographics: 
\code{\link{ethnicn}()},
\code{\link{is_american_native}()},
\code{\link{is_asian}()},
\code{\link{is_black}()},
\code{\link{is_hispanic_or_latino}()},
\code{\link{is_not_hispanic_or_latino}()},
\code{\link{is_other}()},
\code{\link{is_pacific_islander}()},
\code{\link{is_unspecified}()},
\code{\link{is_white}()},
\code{\link{racen}()},
\code{\link{sexf}()}
}
\concept{demographics}
