% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_female}
\alias{is_female}
\title{Takes character input and returns TRUE/FALSE if female/male}
\usage{
is_female(x)
}
\arguments{
\item{x}{input character representing female or male}
}
\value{
boolean representing female
}
\description{
Takes character input and returns TRUE/FALSE if female/male
}
\examples{
is_female("F")

is_female(c("MALE", "FEMALE"))
}
