% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_hispanic_or_latino}
\alias{is_hispanic_or_latino}
\title{Check if Ethnicity is Hispanic or Latino}
\usage{
is_hispanic_or_latino(x)
}
\arguments{
\item{x}{input character representing ethnicity}
}
\value{
boolean representing Ethnic == "Hispanic or Latino"
}
\description{
Check if Ethnicity is Hispanic or Latino
}
\examples{
is_hispanic_or_latino("HISPANIC OR LATINO")

is_hispanic_or_latino("NOT HISPANIC OR LATINO")

is_hispanic_or_latino("UNKNOWN")

is_hispanic_or_latino(1)
}
\seealso{
Other demographics: 
\code{\link{ethnicn}()},
\code{\link{is_asian}()},
\code{\link{is_black}()},
\code{\link{is_female}()},
\code{\link{is_not_hispanic_or_latino}()},
\code{\link{is_other}()},
\code{\link{is_white}()},
\code{\link{racen}()},
\code{\link{sexf}()}
}
\concept{demographics}
