% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_hispanic_or_latino}
\alias{is_hispanic_or_latino}
\title{Takes character input and returns TRUE/FALSE if "Hispanic or Latino" or other}
\usage{
is_hispanic_or_latino(x)
}
\arguments{
\item{x}{input character representing ethnicity}
}
\value{
boolean representing Ethnic == "Hispanic or Latino"
}
\description{
Takes character input and returns TRUE/FALSE if "Hispanic or Latino" or other
}
\examples{
is_hispanic_or_latino("HISPANIC OR LATINO")

is_hispanic_or_latino("NOT HISPANIC OR LATINO")

is_hispanic_or_latino("UNKNOWN")
}
