% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_not_hispanic_or_latino}
\alias{is_not_hispanic_or_latino}
\title{Check if Ethnicity is Not Hispanic or Latino}
\usage{
is_not_hispanic_or_latino(x)
}
\arguments{
\item{x}{input character representing ethnicity}
}
\value{
boolean representing Ethnic == "Not Hispanic or Latino"
}
\description{
Check if Ethnicity is Not Hispanic or Latino
}
\examples{
is_not_hispanic_or_latino("HISPANIC OR LATINO")

is_not_hispanic_or_latino("NOT HISPANIC OR LATINO")

is_not_hispanic_or_latino("UNKNOWN")

is_not_hispanic_or_latino(0)
}
\seealso{
Other demographics:
\code{\link[=ethnicn]{ethnicn()}},
\code{\link[=is_american_native]{is_american_native()}},
\code{\link[=is_asian]{is_asian()}},
\code{\link[=is_black]{is_black()}},
\code{\link[=is_female]{is_female()}},
\code{\link[=is_hispanic_or_latino]{is_hispanic_or_latino()}},
\code{\link[=is_other]{is_other()}},
\code{\link[=is_pacific_islander]{is_pacific_islander()}},
\code{\link[=is_unspecified]{is_unspecified()}},
\code{\link[=is_white]{is_white()}},
\code{\link[=racen]{racen()}},
\code{\link[=sexf]{sexf()}}
}
\concept{demographics}
