% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_other}
\alias{is_other}
\title{Check if Race is Other}
\usage{
is_other(x)
}
\arguments{
\item{x}{input character representing race}
}
\value{
boolean representing Race == Other
}
\description{
Check if Race is Other
}
\examples{
is_other("OTHER")

is_other("MULTIPLE")

is_other("BLACK")

is_other(6)
}
\seealso{
Other demographics: 
\code{\link{ethnicn}()},
\code{\link{is_american_native}()},
\code{\link{is_asian}()},
\code{\link{is_black}()},
\code{\link{is_female}()},
\code{\link{is_hispanic_or_latino}()},
\code{\link{is_not_hispanic_or_latino}()},
\code{\link{is_pacific_islander}()},
\code{\link{is_unspecified}()},
\code{\link{is_white}()},
\code{\link{racen}()},
\code{\link{sexf}()}
}
\concept{demographics}
