% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_other}
\alias{is_other}
\title{Takes character input and returns TRUE/FALSE if other/explicit race}
\usage{
is_other(x)
}
\arguments{
\item{x}{input character representing race}
}
\value{
boolean representing Race == Other
}
\description{
Takes character input and returns TRUE/FALSE if other/explicit race
}
\examples{
is_other("OTHER")

is_other("BLACK")
}
