% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_pacific_islander}
\alias{is_pacific_islander}
\title{Check if Race is Pacific Islander}
\usage{
is_pacific_islander(x)
}
\arguments{
\item{x}{input character representing race}
}
\value{
boolean representing Race == Pacific Islander
}
\description{
Check if Race is Pacific Islander
}
\examples{
is_pacific_islander("NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER")

is_pacific_islander("BLACK")

is_pacific_islander(5)
}
\seealso{
Other demographics:
\code{\link[=ethnicn]{ethnicn()}},
\code{\link[=is_american_native]{is_american_native()}},
\code{\link[=is_asian]{is_asian()}},
\code{\link[=is_black]{is_black()}},
\code{\link[=is_female]{is_female()}},
\code{\link[=is_hispanic_or_latino]{is_hispanic_or_latino()}},
\code{\link[=is_not_hispanic_or_latino]{is_not_hispanic_or_latino()}},
\code{\link[=is_other]{is_other()}},
\code{\link[=is_unspecified]{is_unspecified()}},
\code{\link[=is_white]{is_white()}},
\code{\link[=racen]{racen()}},
\code{\link[=sexf]{sexf()}}
}
\concept{demographics}
