% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_white}
\alias{is_white}
\title{Check if Race is White}
\usage{
is_white(x)
}
\arguments{
\item{x}{input character representing race}
}
\value{
boolean representing Race == White
}
\description{
Check if Race is White
}
\examples{
is_white("WHITE")

is_white("BLACK")

is_white(1)
}
\seealso{
Other demographics:
\code{\link[=ethnicn]{ethnicn()}},
\code{\link[=is_american_native]{is_american_native()}},
\code{\link[=is_asian]{is_asian()}},
\code{\link[=is_black]{is_black()}},
\code{\link[=is_female]{is_female()}},
\code{\link[=is_hispanic_or_latino]{is_hispanic_or_latino()}},
\code{\link[=is_not_hispanic_or_latino]{is_not_hispanic_or_latino()}},
\code{\link[=is_other]{is_other()}},
\code{\link[=is_pacific_islander]{is_pacific_islander()}},
\code{\link[=is_unspecified]{is_unspecified()}},
\code{\link[=racen]{racen()}},
\code{\link[=sexf]{sexf()}}
}
\concept{demographics}
