% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_white}
\alias{is_white}
\title{Takes character input and returns TRUE/FALSE if white/other}
\usage{
is_white(x)
}
\arguments{
\item{x}{input character representing race}
}
\value{
boolean representing Race == White
}
\description{
Takes character input and returns TRUE/FALSE if white/other
}
\examples{
is_white("WHITE")

is_white("BLACK")
}
