% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{mdrd_egfr}
\alias{mdrd_egfr}
\title{Modification of Diet in Renal Disease eGFR calculation}
\usage{
mdrd_egfr(sexf, raceb, age, creat)
}
\arguments{
\item{sexf}{a boolean representing if the patient is female.}

\item{raceb}{a boolean representing if the patient is black.}

\item{age}{the age of the patient in years}

\item{creat}{the serum creatinine levels in mg/dL}
}
\value{
the eGFR in mL/min/1.73 m^2
}
\description{
Modification of Diet in Renal Disease eGFR calculation
}
\examples{
e <- mdrd_egfr(TRUE, TRUE, 24, 1)

df <- data.frame(
   "SEXF" = c(TRUE, FALSE, TRUE, FALSE),
   "RACEB" = c(FALSE, FALSE, TRUE, FALSE),
   "AGE" = c(24, 24, 23, 24),
   "CREAT" = c(1, 1, 2, 1)
   )
df <- dplyr::mutate(df, egfr = mdrd_egfr(SEXF, RACEB, AGE, CREAT))
}
