% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{mosteller_bsa}
\alias{mosteller_bsa}
\title{Calculate Body Surface Area Using Mosteller Equation}
\usage{
mosteller_bsa(weight, height)
}
\arguments{
\item{weight}{weight of subject (kg)}

\item{height}{height of subject (cm)}
}
\value{
the body surface area (m^2)
}
\description{
Calculate Body Surface Area Using Mosteller Equation
}
\details{
The Mosteller equation for BSA:
\deqn{BSA = \sqrt{\frac{W \cdot H}{3600}}}{BSA = sqrt(W * H / 3600)}

where:
\itemize{
\item \eqn{W} = weight (kg)
\item \eqn{H} = height (cm)
}
}
\examples{
mosteller_bsa(70, 170)
}
\seealso{
Other body_composition: 
\code{\link{aibw}()},
\code{\link{bmi}()},
\code{\link{bmic}()},
\code{\link{bsa}()},
\code{\link{dubois_bsa}()},
\code{\link{ibw}()}
}
\concept{body_composition}
