% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{mosteller_bsa}
\alias{mosteller_bsa}
\title{Calculates Body Surface Area based on Weight and Height using Mosteller equation}
\usage{
mosteller_bsa(weight, height)
}
\arguments{
\item{weight}{weight of subject (kg)}

\item{height}{height of subject (cm)}
}
\value{
the body surface area (m^2)
}
\description{
Calculates Body Surface Area based on Weight and Height using Mosteller equation
}
\examples{
mosteller_bsa(70, 170)
}
