% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yspec_def.R
\name{racen}
\alias{racen}
\title{Convert Race to Numeric Code}
\usage{
racen(racec)
}
\arguments{
\item{racec}{Race character}
}
\value{
the standard yspec numeric value for the inputted Race character
}
\description{
Convert Race to Numeric Code
}
\examples{
racen("WHITE") # 1

racen("BLACK") # 2

racen("ASIAN") # 3

racen("OTHER") # 4

racen("UNKNOWN") # default missing value
}
\seealso{
Other demographics: 
\code{\link{ethnicn}()},
\code{\link{is_asian}()},
\code{\link{is_black}()},
\code{\link{is_female}()},
\code{\link{is_hispanic_or_latino}()},
\code{\link{is_not_hispanic_or_latino}()},
\code{\link{is_other}()},
\code{\link{is_white}()},
\code{\link{sexf}()}
}
\concept{demographics}
