% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yspec_def.R
\name{racen}
\alias{racen}
\title{Takes character input and returns standard yspec numeric value for Race}
\usage{
racen(racec)
}
\arguments{
\item{racec}{Race character}
}
\value{
the standard yspec numeric value for the inputted Race character
}
\description{
Takes character input and returns standard yspec numeric value for Race
}
\examples{
racen("WHITE") # 1

racen("BLACK") # 2

racen("ASIAN") # 3

racen("OTHER") # 4

racen("UNKNOWN") # -999
}
