% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_with_hash.R
\name{read_csv_with_hash}
\alias{read_csv_with_hash}
\title{Read CSV File with Hash Verification}
\usage{
read_csv_with_hash(csv_file_path, ..., algo = "blake3")
}
\arguments{
\item{csv_file_path}{path to csv file to ingest}

\item{...}{additional arguments for digest or read_csv}

\item{algo}{hashing algorithm to use, default is "blake3"}
}
\value{
dataframe of data within file
}
\description{
Read CSV File with Hash Verification
}
\examples{
\dontrun{
read_csv_with_hash("data/derived/example_data.csv")
}
}
\seealso{
Other file_io: 
\code{\link{read_excel_with_hash}()},
\code{\link{read_file_with_hash}()},
\code{\link{read_hashed_file}()},
\code{\link{read_parquet_with_hash}()},
\code{\link{read_pzfx_with_hash}()},
\code{\link{read_sas_with_hash}()},
\code{\link{read_xpt_with_hash}()},
\code{\link{write_csv_with_hash}()},
\code{\link{write_file_with_hash}()},
\code{\link{write_parquet_with_hash}()}
}
\concept{file_io}
