% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_with_hash.R
\name{read_csv_with_hash}
\alias{read_csv_with_hash}
\title{Reads data from csv file and prints hash of contents.}
\usage{
read_csv_with_hash(csv_file_path, ...)
}
\arguments{
\item{csv_file_path}{path to csv file to ingest}

\item{...}{additional arguments for digest or read_csv}
}
\value{
dataframe of data within file
}
\description{
Reads data from csv file and prints hash of contents.
}
\examples{
\dontrun{
read_csv_with_hash("data/derived/example_data.csv")
}
}
