% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_with_hash.R
\name{read_excel_with_hash}
\alias{read_excel_with_hash}
\title{Reads data from xlsx/xls file and prints hash of contents.}
\usage{
read_excel_with_hash(xlsx_file_path, ...)
}
\arguments{
\item{xlsx_file_path}{an xlsx/xls file to ingest}

\item{...}{additional arguments to digest or read_excel}
}
\value{
a dataframe(?) of data within file
}
\description{
Reads data from xlsx/xls file and prints hash of contents.
}
\examples{
\dontrun{
read_excel_with_hash("data/source/example.xpt")
}
}
