% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_with_hash.R
\name{read_file_with_hash}
\alias{read_file_with_hash}
\title{Read Data File with Hash Verification}
\usage{
read_file_with_hash(file_path, ..., algo = "blake3")
}
\arguments{
\item{file_path}{path to data file}

\item{...}{additional arguments to digest, read_csv, read_parquet, read_sas, read_pzfx, read_xpt}

\item{algo}{hashing algorithm to use, default is "blake3"}
}
\value{
data within the supplied file
}
\description{
Read Data File with Hash Verification
}
\examples{
\dontrun{
dat <- read_file_with_hash("data/derived/PK_data.parquet")
dat2 <- read_file_with_hash("data/source/data.csv")
}
}
\seealso{
Other file_io: 
\code{\link{read_csv_with_hash}()},
\code{\link{read_excel_with_hash}()},
\code{\link{read_hashed_file}()},
\code{\link{read_parquet_with_hash}()},
\code{\link{read_pzfx_with_hash}()},
\code{\link{read_sas_with_hash}()},
\code{\link{read_xpt_with_hash}()},
\code{\link{write_csv_with_hash}()},
\code{\link{write_file_with_hash}()},
\code{\link{write_parquet_with_hash}()}
}
\concept{file_io}
