% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_with_hash.R
\name{read_file_with_hash}
\alias{read_file_with_hash}
\title{Reads the data from a file (csv or parquet) and prints the hash}
\usage{
read_file_with_hash(file_path, ...)
}
\arguments{
\item{file_path}{path to data file}

\item{...}{additional arguments to digest, read_csv, read_parquet, read_sas, read_pzfx}
}
\value{
data within the supplied file
}
\description{
Reads the data from a file (csv or parquet) and prints the hash
}
\examples{
\dontrun{
dat <- read_file_with_hash("data/derived/PK_data.parquet")
dat2 <- read_file_with_hash("data/source/data.csv")
}
}
