% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_with_hash.R
\name{read_file_with_hash}
\alias{read_file_with_hash}
\title{Reads the data from a file (csv or parquet) and prints the hash}
\usage{
read_file_with_hash(
  file_path,
  na = c("", "NA"),
  show_col_types = TRUE,
  table = NULL
)
}
\arguments{
\item{file_path}{path to data file}

\item{na}{Character vector of strings to interpret as missing values.}

\item{show_col_types}{logical values to show column types when reading csv files.}

\item{table}{optional argument only needed for reading prism pzfx files.}
}
\value{
data within the supplied file
}
\description{
Reads the data from a file (csv or parquet) and prints the hash
}
\examples{
\dontrun{
dat <- read_file_with_hash("data/derived/PK_data.parquet")
dat2 <- read_file_with_hash("data/source/data.csv")
}
}
