% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_with_hash.R
\name{read_hashed_file}
\alias{read_hashed_file}
\title{Read File with Required Hash Match}
\usage{
read_hashed_file(file_path, hash, ..., algo = "blake3")
}
\arguments{
\item{file_path}{path to file with data you want to read}

\item{hash}{hash you expect the file to have}

\item{...}{additional arguments for digest or read_csv, parquet, sas}

\item{algo}{hashing algorithm to use, default is "blake3"}
}
\value{
data object of contents of file_path
}
\description{
Read File with Required Hash Match
}
\examples{
\dontrun{
file_path <- "data/derived/example_pk.parquet"

hash <- 0cfd6da55e6c1e198effe1e584c26d79
read_hashed_file(file_path, hash)
}
}
\seealso{
Other file_io: 
\code{\link{read_csv_with_hash}()},
\code{\link{read_excel_with_hash}()},
\code{\link{read_file_with_hash}()},
\code{\link{read_parquet_with_hash}()},
\code{\link{read_pzfx_with_hash}()},
\code{\link{read_sas_with_hash}()},
\code{\link{read_xpt_with_hash}()},
\code{\link{write_csv_with_hash}()},
\code{\link{write_file_with_hash}()},
\code{\link{write_parquet_with_hash}()}
}
\concept{file_io}
