% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_with_hash.R
\name{read_hashed_file}
\alias{read_hashed_file}
\title{Reads a file if the supplied hash matches the file's hash}
\usage{
read_hashed_file(file_path, hash)
}
\arguments{
\item{file_path}{path to file with data you want to read}

\item{hash}{hash you expect the file to have}
}
\value{
data object of contents of file_path
}
\description{
Reads a file if the supplied hash matches the file's hash
}
\examples{
\dontrun{
file_path <- "data/derived/example_pk.parquet"

hash <- 0cfd6da55e6c1e198effe1e584c26d79
read_hashed_file(file_path, hash)
}
}
