% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_with_hash.R
\name{read_parquet_with_hash}
\alias{read_parquet_with_hash}
\title{Read Parquet File with Hash Verification}
\usage{
read_parquet_with_hash(parquet_file_path, ..., algo = "blake3")
}
\arguments{
\item{parquet_file_path}{path to parquet file to ingest}

\item{...}{additional arguments to digest or read_parquet}

\item{algo}{hashing algorithm to use, default is "blake3"}
}
\value{
a tibble of data within file
}
\description{
Read Parquet File with Hash Verification
}
\examples{
\dontrun{
read_parquet_with_hash("data/derived/example_data.parquet")
}
}
\seealso{
Other file_io: 
\code{\link{read_csv_with_hash}()},
\code{\link{read_excel_with_hash}()},
\code{\link{read_file_with_hash}()},
\code{\link{read_hashed_file}()},
\code{\link{read_pzfx_with_hash}()},
\code{\link{read_sas_with_hash}()},
\code{\link{read_xpt_with_hash}()},
\code{\link{write_csv_with_hash}()},
\code{\link{write_file_with_hash}()},
\code{\link{write_parquet_with_hash}()}
}
\concept{file_io}
