% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_with_hash.R
\name{read_pzfx_with_hash}
\alias{read_pzfx_with_hash}
\title{Reads in table from a prism pzfx file.}
\usage{
read_pzfx_with_hash(pzfx_file_path, table)
}
\arguments{
\item{pzfx_file_path}{path to pzfx file}

\item{table}{name of table wihtin file to read. Must be in pzfx_tables(pzfx_file_path)}
}
\value{
data within the table of the pzfx file
}
\description{
Reads in table from a prism pzfx file.
}
\examples{
\dontrun{
read_pzfx_with_hash("mydata.pzfx", table = "experiment1")
}
}
