% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_with_hash.R
\name{read_sas_with_hash}
\alias{read_sas_with_hash}
\title{Reads data from sas file and prints hash of contents.}
\usage{
read_sas_with_hash(sas_file_path)
}
\arguments{
\item{sas_file_path}{path to sas file to ingest}
}
\value{
a dataframe(?) of data within file
}
\description{
Reads data from sas file and prints hash of contents.
}
\examples{
\dontrun{
read_sas_with_hash("data/source/example.sas7bdat")
}
}
