% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{schwartz_egfr}
\alias{schwartz_egfr}
\title{Calculates eGFR based on Schwartz' equation}
\usage{
schwartz_egfr(height, creat)
}
\arguments{
\item{height}{height of patients in cm.}

\item{creat}{Serum creatinine levels in mg/dL}
}
\value{
eGFR in mL/min/1.73m^2
}
\description{
Calculates eGFR based on Schwartz' equation
}
\examples{
schwartz_egfr(100, 1)
}
