% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scicalc-package.R
\docType{package}
\name{scicalc-package}
\alias{scicalc}
\alias{scicalc-package}
\title{scicalc: Scientific Calculations for Quantitative Clinical Pharmacology and Pharmacometrics Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utility functions helpful for reproducible scientific calculations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://a2-ai.github.io/scicalc-docs}
  \item \url{https://github.com/a2-ai/scicalc}
}

}
\author{
\strong{Maintainer}: Matthew Smith \email{matthews@a2-ai.com}

Authors:
\itemize{
  \item Jacob Dumbleton \email{jacob@a2-ai.com}
  \item Jenna Johnson \email{jenna@a2-ai.com}
  \item Devin Pastoor \email{devin@a2-ai.com}
}

Other contributors:
\itemize{
  \item Wesley Cummings \email{wes@a2-ai.com} [contributor]
  \item Emily Schapiro \email{emily@a2-ai.com} [contributor]
  \item Ryan Crass \email{ryan@a2-ai.com} [contributor]
  \item Jonah Lyon \email{jonah@a2-ai.com} [contributor]
  \item Elizabeth LeBeau \email{elizabeth@a2-ai.com} [contributor]
}

}
\keyword{internal}
