% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yspec_def.R
\name{sexf}
\alias{sexf}
\title{Convert Sex to Numeric Code}
\usage{
sexf(sex)
}
\arguments{
\item{sex}{Sex character}
}
\value{
the standard yspec numeric value for the inputted Sex character
}
\description{
Also returns numeric for single character Sex characters "F" and "M"
}
\examples{
sexf("FEMALE") # 1
sexf("female") # 1
sexf("f") # 1

sexf("MALE") # 0

sexf("NOT SPECIFIED") # 0
}
\seealso{
Other demographics:
\code{\link[=ethnicn]{ethnicn()}},
\code{\link[=is_american_native]{is_american_native()}},
\code{\link[=is_asian]{is_asian()}},
\code{\link[=is_black]{is_black()}},
\code{\link[=is_female]{is_female()}},
\code{\link[=is_hispanic_or_latino]{is_hispanic_or_latino()}},
\code{\link[=is_not_hispanic_or_latino]{is_not_hispanic_or_latino()}},
\code{\link[=is_other]{is_other()}},
\code{\link[=is_pacific_islander]{is_pacific_islander()}},
\code{\link[=is_unspecified]{is_unspecified()}},
\code{\link[=is_white]{is_white()}},
\code{\link[=racen]{racen()}}
}
\concept{demographics}
