% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yspec_def.R
\name{sexf}
\alias{sexf}
\title{Takes character input and returns standard yspec numeric value for Sex.}
\usage{
sexf(sex)
}
\arguments{
\item{sex}{Sex character}
}
\value{
the standard yspec numeric value for the inputted Sex character
}
\description{
Also returns numeric for single character Sex characters "F" and "M"
}
\examples{
sexf("FEMALE") # 1
sexf("female") # 1
sexf("f") # 1

sexf("MALE") # 0

sexf("NOT SPECIFIED") # 0
}
