% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_csv_with_hash}
\alias{write_csv_with_hash}
\title{Write CSV File with Hash Output}
\usage{
write_csv_with_hash(data, csv_path, ...)
}
\arguments{
\item{data}{a data object to write to file}

\item{csv_path}{the file path to save the csv}

\item{...}{additional arguments to digest or write_csv}
}
\value{
Nothing, creates csv_path file and prints hash of the file
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated; use \code{\link[=write_file_with_hash]{write_file_with_hash()}}.
}
\examples{
\dontrun{
df <- data.frame(
  "a" = c(1, 2, 3, 4)
  "b" = c("A", "B", "C", "D")
)
write_file_with_hash(df, "test/test.csv")
}
}
\keyword{internal}
