% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_csv_with_hash}
\alias{write_csv_with_hash}
\title{Write CSV File with Hash Output}
\usage{
write_csv_with_hash(data, csv_path, ...)
}
\arguments{
\item{data}{a data object to write to file}

\item{csv_path}{the file path to save the csv}

\item{...}{additional arguments to digest or write_csv}
}
\value{
Nothing, creates csv_path file and prints hash of the file
}
\description{
Write CSV File with Hash Output
}
\examples{
\dontrun{
df <- data.frame(
  "a" = c(1, 2, 3, 4)
  "b" = c("A", "B", "C", "D")
)
write_csv_with_hash(df, "test/test.csv")
}
}
\seealso{
Other file_io: 
\code{\link{read_csv_with_hash}()},
\code{\link{read_excel_with_hash}()},
\code{\link{read_file_with_hash}()},
\code{\link{read_hashed_file}()},
\code{\link{read_parquet_with_hash}()},
\code{\link{read_pzfx_with_hash}()},
\code{\link{read_sas_with_hash}()},
\code{\link{read_xpt_with_hash}()},
\code{\link{write_file_with_hash}()},
\code{\link{write_parquet_with_hash}()}
}
\concept{file_io}
