% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_csv_with_hash}
\alias{write_csv_with_hash}
\title{Writes data to csv_path with na_value replacing NA values.}
\usage{
write_csv_with_hash(data, csv_path, ...)
}
\arguments{
\item{data}{a data object to write to file}

\item{csv_path}{the file path to save the csv}

\item{...}{additional arguments to digest or write_csv}
}
\value{
Nothing, creates csv_path file and prints hash of the file
}
\description{
Writes data to csv_path with na_value replacing NA values.
}
\examples{
\dontrun{
df <- data.frame(
  "a" = c(1, 2, 3, 4)
  "b" = c("A", "B", "C", "D")
)
write_csv_with_hash(df, "test/test.csv")
}
}
