% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_file_with_hash}
\alias{write_file_with_hash}
\title{Write Data File with Hash Output}
\usage{
write_file_with_hash(data, path, overwrite = FALSE, ...)
}
\arguments{
\item{data}{the data object to write to file}

\item{path}{the destination of the file (csv or parquet)}

\item{overwrite}{boolean of whether to overwrite or not.}

\item{...}{additional arguments for digest or write_file.}
}
\value{
Nothing, File is created and hash of created file is printed
}
\description{
Write Data File with Hash Output
}
\examples{
\dontrun{
df <- data.frame(
  "a" = c(1, 2, 3, 4)
  "b" = c("A", "B", "C", "D")
)
write_data_with_hash(df, "data.csv")
}
}
\seealso{
Other file_io: 
\code{\link{read_csv_with_hash}()},
\code{\link{read_excel_with_hash}()},
\code{\link{read_file_with_hash}()},
\code{\link{read_hashed_file}()},
\code{\link{read_parquet_with_hash}()},
\code{\link{read_pzfx_with_hash}()},
\code{\link{read_sas_with_hash}()},
\code{\link{read_xpt_with_hash}()},
\code{\link{write_csv_with_hash}()},
\code{\link{write_parquet_with_hash}()}
}
\concept{file_io}
