% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_file_with_hash}
\alias{write_file_with_hash}
\title{Writes data to path, if directory doesn't exist it is created before file is written}
\usage{
write_file_with_hash(data, path, overwrite = FALSE, na_value = ".")
}
\arguments{
\item{data}{the data object to write to file}

\item{path}{the destination of the file (csv or parquet)}

\item{overwrite}{boolean of whether to overwrite or not.}

\item{na_value}{string for what to replace NA with. default is "."}
}
\value{
Nothing, File is created and hash of created file is printed
}
\description{
Writes data to path, if directory doesn't exist it is created before file is written
}
\examples{
\dontrun{
df <- data.frame(
  "a" = c(1, 2, 3, 4)
  "b" = c("A", "B", "C", "D")
)
write_data_with_hash(df, "data.csv")
}
}
