% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_parquet_with_hash}
\alias{write_parquet_with_hash}
\title{Write Parquet File with Hash Output}
\usage{
write_parquet_with_hash(data, parquet_path, ...)
}
\arguments{
\item{data}{the data object to save to parquet_path}

\item{parquet_path}{the path to the desired parquet destination}

\item{...}{additional arguments to digest and write_parquet}
}
\value{
Nothing. creates parquet_path file and prints hash
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated; use \code{\link[=write_file_with_hash]{write_file_with_hash()}}.
}
\examples{
\dontrun{
df <- data.frame(
  "a" = c(1, 2, 3, 4)
  "b" = c("A", "B", "C", "D")
)
write_file_with_hash(df, "test/test.parquet")
}
}
\seealso{
Other file_io: 
\code{\link{read_csv_with_hash}()},
\code{\link{read_excel_with_hash}()},
\code{\link{read_file_with_hash}()},
\code{\link{read_hashed_file}()},
\code{\link{read_parquet_with_hash}()},
\code{\link{read_pzfx_with_hash}()},
\code{\link{read_sas_with_hash}()},
\code{\link{read_xpt_with_hash}()},
\code{\link{write_csv_with_hash}()},
\code{\link{write_file_with_hash}()}
}
\concept{file_io}
\keyword{internal}
