% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_parquet_with_hash}
\alias{write_parquet_with_hash}
\title{Writes data to parquet_path and prints hash}
\usage{
write_parquet_with_hash(data, parquet_path)
}
\arguments{
\item{data}{the data object to save to parquet_path}

\item{parquet_path}{the path to the desired parquet destination}
}
\value{
Nothing. creates parquet_path file and prints hash
}
\description{
Writes data to parquet_path and prints hash
}
\examples{
\dontrun{
df <- data.frame(
  "a" = c(1, 2, 3, 4)
  "b" = c("A", "B", "C", "D")
)
write_parquet_with_hash(df, "test/test.parquet")
}
}
